+++ src/lpj/fopenoutput.c
@@ -25,13 +25,13 @@ static Bool create(Netcdf *cdf,const char *filename,int index,
                   config->npft[CROP],config);
   if(size==1)
     return create_netcdf(cdf,filename,
-                         (config->outputvars[index].id==GRID) ? "soilcode" :
+                         (config->outputvars[index].id==GRID) ? "cellid" :
                          config->outnames[config->outputvars[index].id].var,
-                         (config->outputvars[index].id==GRID) ? "soil code" :
+                         (config->outputvars[index].id==GRID) ? "cell id" :
                          config->outnames[config->outputvars[index].id].descr,
                          (config->outputvars[index].id==GRID) ? "" :
                          config->outnames[config->outputvars[index].id].unit,
-                         getoutputtype(config->outputvars[index].id,FALSE),
+                         (config->outputvars[index].id==GRID) ? LPJ_INT : getoutputtype(config->outputvars[index].id,FALSE),
                          getnyear(config->outnames,config->outputvars[index].id),
                          (config->outnames[config->outputvars[index].id].timestep==ANNUAL) ? 1 : config->outnames[config->outputvars[index].id].timestep,array,config);
   else
@@ -78,7 +78,7 @@ static void openfile(Outputfile *output,const Cell grid[],
           }
         }
         break;
-    }
+    } /* of switch */
     if(isroot(*config))
     {
       if(!config->outputvars[i].oneyear)
@@ -204,7 +204,7 @@ static void openfile(Outputfile *output,const Cell grid[],
         else
           output->files[config->outputvars[i].id].isopen=TRUE;
         break;
-    }
+    } /* of switch */
   }
   output->files[config->outputvars[i].id].oneyear=config->outputvars[i].oneyear;
 } /* of 'openfile' */
@@ -302,51 +302,60 @@ Outputfile *fopenoutput(const Cell grid[],   /**< LPJ grid */
 
 void openoutput_yearly(Outputfile *output,int year,const Config *config)
 {
-  String filename;
+  char *filename;
   Header header;
-  int i,size;
+  int i,size,count;
   for(i=0;i<config->n_out;i++)
     if(config->outputvars[i].oneyear)
     {
       if(isroot(*config))
       {
-        snprintf(filename,STRING_LEN,config->outputvars[i].filename.name,year);
-        if(config->outputvars[i].filename.meta)
-          fprintoutputjson(i,year,config);
-        switch(config->outputvars[i].filename.fmt)
+        count=snprintf(NULL,0,config->outputvars[i].filename.name,year);
+        if(count==-1)
         {
-          case CLM:
-            if((output->files[config->outputvars[i].id].fp.file=fopen(filename,"wb"))==NULL)
-            {
-              printfcreateerr(filename);
-              output->files[config->outputvars[i].id].isopen=FALSE;
-            }
-            else
-            {
-              output->files[config->outputvars[i].id].isopen=TRUE;
-              header.firstyear=year;
-              header.nyear=1;
-              header.ncell=config->total;
-              header.firstcell=config->firstgrid;
-              header.cellsize_lon=(float)config->resolution.lon;
-              header.cellsize_lat=(float)config->resolution.lat;
-              header.scalar=1;
-              header.order=CELLSEQ;
-              header.timestep=1;
-              header.nstep=getnyear(config->outnames,config->outputvars[i].id);
-              header.nbands=outputsize(config->outputvars[i].id,
-                                       config->npft[GRASS]+config->npft[TREE],
-                                       config->npft[CROP],config);
-              if(config->outputvars[i].id==SDATE || config->outputvars[i].id==HDATE || config->outputvars[i].id==SEASONALITY)
-                header.datatype=LPJ_SHORT;
+          fprintf(stderr,"ERROR247: Invalid format in filename '%s'.\n",config->outputvars[i].filename.name);
+          output->files[config->outputvars[i].id].isopen=FALSE;
+        }
+        else
+        {
+          filename=malloc(count+1);
+          check(filename);
+          snprintf(filename,count+1,config->outputvars[i].filename.name,year);
+          if(config->outputvars[i].filename.meta)
+            fprintoutputjson(i,year,config);
+          switch(config->outputvars[i].filename.fmt)
+          {
+            case CLM:
+              if((output->files[config->outputvars[i].id].fp.file=fopen(filename,"wb"))==NULL)
+              {
+                printfcreateerr(filename);
+                output->files[config->outputvars[i].id].isopen=FALSE;
+              }
               else
-                header.datatype=LPJ_FLOAT;
-              fwriteheader(output->files[config->outputvars[i].id].fp.file,
-                           &header,LPJOUTPUT_HEADER,config->outputvars[i].filename.version);
+              {
+                output->files[config->outputvars[i].id].isopen=TRUE;
+                header.firstyear=year;
+                header.nyear=1;
+                header.ncell=config->total;
+                header.firstcell=config->firstgrid;
+                header.cellsize_lon=(float)config->resolution.lon;
+                header.cellsize_lat=(float)config->resolution.lat;
+                header.scalar=1;
+                header.order=CELLSEQ;
+                header.timestep=1;
+                header.nstep=getnyear(config->outnames,config->outputvars[i].id);
+                header.nbands=outputsize(config->outputvars[i].id,
+                                         config->npft[GRASS]+config->npft[TREE],
+                                         config->npft[CROP],config);
+                if(config->outputvars[i].id==SDATE || config->outputvars[i].id==HDATE || config->outputvars[i].id==SEASONALITY)
+                  header.datatype=LPJ_SHORT;
+                else
+                  header.datatype=LPJ_FLOAT;
+                fwriteheader(output->files[config->outputvars[i].id].fp.file,
+                             &header,LPJOUTPUT_HEADER,config->outputvars[i].filename.version);
 
-            }
-            break;
-
+              }
+              break;
           case RAW:
             if((output->files[config->outputvars[i].id].fp.file=fopen(filename,"wb"))==NULL)
             {
@@ -388,8 +397,10 @@ void openoutput_yearly(Outputfile *output,int year,const Config *config)
                            getoutputtype(config->outputvars[i].id,FALSE),
                            getnyear(config->outnames,config->outputvars[i].id),year,output->index,config);
 
+          } /* of switch */
+          free(filename);
         }
-     }
+     } /* of(isroot(*config)) */
 #ifdef USE_MPI
      MPI_Bcast(&output->files[config->outputvars[i].id].isopen,1,MPI_INT,0,config->comm);
 #endif